refactor: review round 1 should-fix findings
- Extract useDraggable hook from color-tweak-panel and demo-tweak-panel,
  fixing useCallback [position] dependency causing recreation every frame
- Switch overlay panels from client:load to client:idle for lazy hydration
- Extract shared jsonResponse into src/lib/api-utils.ts (DRY)
- Memoize generateBaseTokensCss() in HtmlPreview to avoid recomputation
- Add unit tests to b4push script and CI workflow
- Use DEMO_PANEL_WIDTH constant instead of hardcoded 280
