+++ .github/workflows/ci.yml
@@ -23,6 +23,7 @@ jobs:
           cache: "pnpm"
       - run: pnpm install --frozen-lockfile
       - run: pnpm check
+      - run: pnpm test
       - run: pnpm build
       - name: Doc build
         run: pnpm doc:build
