+++ src/lib/api-utils.ts
@@ -1,0 +1,8 @@
+const JSON_HEADERS = { "Content-Type": "application/json" };
+
+export function jsonResponse(
+  data: Record<string, unknown>,
+  status = 200,
+): Response {
+  return new Response(JSON.stringify(data), { status, headers: JSON_HEADERS });
+}
